refactor family-form into component
